test(generation): split split-robustness into two parallel files
binGenerator.scenario.split-robustness.test.ts (~166s) was the heaviest
single generator test file after the scenario monolith split (#2017), so
it set the floor on the slowest CI shard (~187s, just over 3 min).

Split its 8 describe blocks into two balanced, semantically-grouped files
that run on parallel Vitest workers:
  - split-robustness.test.ts: parameter permutations + regressions
    (wall thickness, base style, lip+wall, connector config, regression)
  - split-robustness-topology.test.ts: cut-topology dimensions
    (interior features, size extremes, split-axis variations)

All 41 tests preserved; run together they drop from ~166s to ~53s wall.
The heaviest generator file is now split-wall-locking (~98s), so the
slowest shard lands well under 3 min.
